Allow dot-completion for attributes
When I write `builtins.`, it wasn't possible until now to list all
attributes in `builtins` (or any other indexed attr-set).

This happened because in that case `utils::ident_at` found `NODE_SELECT`
for the current token (i.e. `TOKEN_DOT`) and not `NODE_IDENT`. This
patch handles the case where `NODE_SELECT` is the current token and
searches for the last `NODE_IDENT`.