Implement first integration test
I think that integration tests are a good idea here to simulate entire
interactions with an LSP (which means we can implement actual
regression-tests in the future when needed) and also check that all the
components work well together.

Unfortunately, I couldn't test *all* layers, but only the
request/notification receivers because I had to replace `App::main` with
a "non-blocking" (well, not quite) variant that can break out of the
loop as soon as the test terminates and the LSP thread is stopped.